backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
boto3==1.28.64
langchain==0.0.335
openai==1.3.0
sentence-transformers==2.2.2
pydantic==2.4.2
python-dotenv==1.0.0
pymupdf==1.23.4 